Update build system to use a more recent version of Laurie's template
Honestly wasn't going to do this, but then I ran into an issue where none of my changes were affecting anything, and I jumped to the conclusion that it was the build process that was acting awry. Turns out the copy script puts the .so in a different folder and was acting as a duplicate to my manually placed .so file. Oh well.